Add initial car motor and ir receiver code
If the ir receiver on the car gets a code that corresponds to one of those sent
by the controller, the motors will move in the appropriate direction.
When no data is received *after a number of ms from the last received
signal*, the car will stop. This timing is necessary to allow the
receiver to *constantly* scan for data without being interrupted between
signals.

One other method for stopping the car is to have the joystick send an ir
code when it is at the origin. When the car receives this code, it can
call a function that turns the motors off. The downside to this is
possible lag when the receiver doesn't receive any data at all (which
can happen quite frequently). In this case, the state the motors were
last in would be kept until another ir code is received. Needs testing to
determine whether this method is better/more reliable than the one used
in this commit.